+++ passes/checks.go
@@ -17,6 +17,7 @@ import (
 	"github.com/bflad/tfproviderlint/passes/R007"
 	"github.com/bflad/tfproviderlint/passes/R008"
 	"github.com/bflad/tfproviderlint/passes/R009"
+	"github.com/bflad/tfproviderlint/passes/R010"
 	"github.com/bflad/tfproviderlint/passes/S001"
 	"github.com/bflad/tfproviderlint/passes/S002"
 	"github.com/bflad/tfproviderlint/passes/S003"
@@ -81,6 +82,7 @@ var AllChecks = []*analysis.Analyzer{
 	R007.Analyzer,
 	R008.Analyzer,
 	R009.Analyzer,
+	R010.Analyzer,
 	S001.Analyzer,
 	S002.Analyzer,
 	S003.Analyzer,