+++ passes/helper/schema/resourcedatagetchangeassignstmt/resourcedatagetchangeassignstmt_test.go
@@ -1,0 +1,15 @@
+package resourcedatagetchangeassignstmt
+
+import (
+	"testing"
+
+	"golang.org/x/tools/go/analysis"
+)
+
+func TestValidateAnalyzer(t *testing.T) {
+	err := analysis.Validate([]*analysis.Analyzer{Analyzer})
+
+	if err != nil {
+		t.Fatal(err)
+	}
+}