+++ CHANGELOG.md
@@ -4,6 +4,7 @@ FEATURES
 
 * **New Check:** `R007`: check for deprecated `(schema.ResourceData).Partial` receiver method usage
 * **New Check:** `R008`: check for deprecated `(schema.ResourceData).SetPartial` receiver method usage
+* **New Check:** `R009`: check for Go panic usage
 
 # v0.9.0